requirements.txt
streamlit>=1.28.0
google-auth-oauthlib>=1.1.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.100.0
gspread>=5.12.0
python-dotenv>=1.0.0
requests>=2.31.0
facebook-sdk>=3.0.0
instagrapi>=2.0.0
passlib>=1.7.4
bcrypt>=4.1.0
PyJWT>=2.8.0
email-validator>=2.1.0